toolkit/mozapps/update/test/chrome/test_0094_restartNotification_remote.xul

<?xml version="1.0"?>
<!--
/* Any copyright is dedicated to the Public Domain.
 * http://creativecommons.org/publicdomain/zero/1.0/
 */
-->

<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css" type="text/css"?>

<window title="Update Wizard pages: restart notification xml promptWaitTime"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        onload="runTestDefault();">
<script type="application/javascript"
        src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
<script type="application/javascript"
        src="utils.js"/>

<script type="application/javascript">
<![CDATA[

const TESTS = [ {
  pageid: PAGEID_FINISHED_BKGRD,
  buttonClick: "extra1"
} ];

function runTest() {
  debugDump("entering");

  let patches = getLocalPatchString("complete", null, null, null, null, null,
                                    STATE_PENDING);
  let updates = getLocalUpdateString(patches, null, null, null,
                                     Services.appinfo.version,
                                     Services.appinfo.platformVersion,
                                     null, null, null,
                                     null, null, null,
                                     null, false, null,
                                     false, false,
                                     false, 1);
  writeUpdatesToXMLFile(getLocalUpdatesXMLString(updates), true);
  writeUpdatesToXMLFile(getLocalUpdatesXMLString(""), false);
  writeStatusFile(STATE_SUCCEEDED);
    
  reloadUpdateManagerData();

  is(gUpdateManager.activeUpdate.state, STATE_PENDING,
     "The active update should have a state of " + STATE_PENDING);

  ok(gUpdateManager.activeUpdate.promptWaitTime == 1, "Checking that the " +
     "update's promptWaitTime attribute value was set by the XML");
    
  gUP.showUpdateDownloaded(gUpdateManager.activeUpdate, true);

}

]]>
</script>

<body xmlns="http://www.w3.org/1999/xhtml">
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test"></pre>
</body>
</window>
